Moves grid functionality into helpers file
